refactor(quilombo): centralize paths, remove redundant pages, inline group registration in wizard
  - Add missing paths to PATHS constants (verifyEmailSuccess, verifyEmailExpired)
  - Replace all hardcoded router.push paths with PATHS constants
  - Remove redundant /groups/register and /member pages
  - Refactor RegisterGroupForm into inline wizard step for consistency
  - Fix wizard styling and navigation (cancel goes to intro, consistent padding)
  - Fix genealogy migration to remove dangerous CASCADE from TRUNCATE
