+++ apps/quilombo/components/forms/group/index.ts
@@ -1,4 +1,3 @@
 export { default as DeleteGroup } from './DeleteGroup';
 export { default as GroupEdit } from './GroupEdit';
 export { default as GroupForm } from './GroupForm';
-export { default as RegisterGroupForm } from './RegisterGroupForm';
