+++ apps/quilombo/components/groups/GroupAssociationWizard.tsx
@@ -6,13 +6,16 @@ import { Building2Icon, CheckCircleIcon, KeyIcon, PlusCircleIcon, SearchIcon, Us
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 import { enqueueSnackbar } from 'notistack';
+import { Field, Form, Formik, type FormikProps } from 'formik';
 
 import GroupSearchSelect from '@/components/genealogy/ui/GroupSearchSelect';
-import { PATHS } from '@/config/constants';
+import { FieldInput, FieldTextarea, StringSelect } from '@/components/forms';
+import { PATHS, styles } from '@/config/constants';
+import { type RegisterGroupForm as RegisterGroupFormType, registerGroupFormSchema } from '@/config/validation-schema';
 import { currentUserAtom } from '@/hooks/state/currentUser';
 import { useFetchUserGroups } from '@/query/currentUser';
 import { useAddRelationship } from '@/query/genealogyProfile';
-import { useGroupClaimable } from '@/query/group';
+import { useGroupClaimable, useRegisterNewGroupMutation } from '@/query/group';
 import ClaimGroupModal from './GroupProfile/ClaimGroupModal';
 
 type GroupProfile = {
@@ -23,7 +26,7 @@ type GroupProfile = {
   isActive: boolean;
 };
 
-type WizardStep = 'intro' | 'search' | 'group-actions' | 'not-found';
+type WizardStep = 'intro' | 'search' | 'group-actions' | 'not-found' | 'register';
 
 /**
  * Wizard for associating a user with a group.
@@ -42,6 +45,8 @@ const GroupAssociationWizard = () => {
   const [wizardStep, setWizardStep] = useState<WizardStep>('intro');
   const [selectedGroup, setSelectedGroup] = useState<GroupProfile | null>(null);
   const [isClaimModalOpen, setIsClaimModalOpen] = useState(false);
+
+  const registerGroupMutation = useRegisterNewGroupMutation();
 
   const addRelationshipMutation = useAddRelationship();
   const { data: userGroups } = useFetchUserGroups();
@@ -341,10 +346,102 @@ const GroupAssociationWizard = () => {
           <Button variant="light" onPress={() => setWizardStep('intro')}>
             Back
           </Button>
-          <Button color="primary" onPress={() => router.push('/groups/register')}>
+          <Button color="primary" onPress={() => setWizardStep('register')}>
             Register New Group
           </Button>
         </CardFooter>
+      </Card>
+    );
+  }
+
+  // Register new group step
+  if (wizardStep === 'register') {
+    const handleSubmit = async (values: RegisterGroupFormType) => {
+      try {
+        await registerGroupMutation.mutateAsync(values);
+        enqueueSnackbar("Registration submitted! We'll review your request and email you.", { variant: 'success' });
+        setWizardStep('intro');
+      } catch (error) {
+        console.error('Error registering group:', error);
+        enqueueSnackbar((error as Error).message || 'An error occurred while submitting your registration', {
+          variant: 'error',
+        });
+      }
+    };
+
+    const initialValues: RegisterGroupFormType = {
+      proposedName: '',
+      proposedStyle: null,
+      website: '',
+      userMessage: '',
+    };
+
+    return (
+      <Card className="max-w-lg">
+        <CardHeader className="flex flex-col items-start gap-1">
+          <h2 className="text-xl font-semibold">Register a New Group</h2>
+          <p className="text-default-500 text-sm">
+            Can&apos;t find your group in our database? Submit a registration request and we&apos;ll review it.
+          </p>
+        </CardHeader>
+        <CardBody className="gap-4">
+          <Formik<RegisterGroupFormType>
+            initialValues={initialValues}
+            validationSchema={registerGroupFormSchema}
+            onSubmit={handleSubmit}
+          >
+            {({ isSubmitting, isValid, dirty }: FormikProps<RegisterGroupFormType>) => (
+              <Form className="flex flex-col gap-4">
+                <Field
+                  name="proposedName"
+                  label="Group Name"
+                  placeholder="e.g., Grupo Capoeira Angola"
+                  as={FieldInput}
+                  isRequired
+                />
+
+                <Field name="proposedStyle" label="Style" options={[...styles]} as={StringSelect} />
+
+                <Field
+                  name="website"
+                  label="Website"
+                  placeholder="https://example.com"
+                  description="Optional - helps us verify your group"
+                  as={FieldInput}
+                />
+
+                <div className="mt-2">
+                  <p className="text-default-600 mb-3 text-sm">To help us verify your registration, please provide:</p>
+                  <ul className="list-disc list-inside text-default-500 text-sm space-y-1 ml-2 mb-3">
+                    <li>Your relationship to the group (founder, leader, instructor, etc.)</li>
+                    <li>Where the group is based</li>
+                    <li>When it was founded (if known)</li>
+                    <li>Any other information that helps verify the group exists</li>
+                  </ul>
+                </div>
+
+                <Field
+                  name="userMessage"
+                  label="Tell us about this group"
+                  placeholder="I am the instructor at this group which was founded in 2010..."
+                  description="Minimum 20 characters required"
+                  as={FieldTextarea}
+                  minRows={4}
+                  isRequired
+                />
+
+                <div className="flex justify-end gap-2 mt-4">
+                  <Button color="default" variant="light" onPress={() => setWizardStep('intro')}>
+                    Cancel
+                  </Button>
+                  <Button color="primary" type="submit" isLoading={isSubmitting} isDisabled={!isValid || !dirty}>
+                    Submit Registration
+                  </Button>
+                </div>
+              </Form>
+            )}
+          </Formik>
+        </CardBody>
       </Card>
     );
   }
